chore: fix merge issue (separate url constants)
