+++ src/components/Main/Recommendations/index.tsx
@@ -1,4 +1,4 @@
-import { PAGE } from '@/constants';
+import { PAGE } from '@/constants/url';
 import { useRecommendations } from '@/hooks/query/useRecommendations';
 import AnimatedTitle from '../_shared/AnimatedTitle';
 import NothingToShow from '../_shared/NothingToShow/NothingToShow';
