+++ src/components/Search/SearchResult/TabContent/_shared/Card/card.module.scss
@@ -1,4 +1,3 @@
-@use '../CoverImage/cover-image';
 @use '@/utils/variables';
 @use '@/utils/mixins';
 @use '../skeleton.scss';
