+++ src/components/Search/SearchResult/TabContent/_shared/PlayButton/play-button.scss
@@ -1,4 +1,3 @@
-@use '../CoverImage/cover-image';
 @use '@/utils/variables';
 
 .play-button {
