+++ src/constants/index.ts
@@ -12,28 +12,6 @@ export const GENIUS_CLIENT_SECRET = import.meta.env.VITE_GENIUS_CLIENT_SECRET;
 export const GENIUS_CLIENT_ACCESS_TOKEN = import.meta.env
   .VITE_GENIUS_CLIENT_ACCESS_TOKEN;
 
-export const API = {
-  LOGIN: 'https://vinylify-express.vercel.app/',
-  EXPRESS: 'https://vinylify-express.vercel.app/',
-  SPOTIFY: 'https://api.spotify.com/v1/',
-  STANDS4: `https://www.stands4.com/services/v2/`,
-  MUSIX_MATCH: `/music-info/api`,
-  GENIUS: '/genius/api',
-};
-export const PAGE = {
-  MAIN: '/',
-  ERROR: '/error',
-  MYPAGE: '/mypage',
-  SEARCH: '/search',
-  MUSIC_INFO: '/music-info',
-  LOGGED_IN: '/me',
-};
-
-export const URL_PARAMS = {
-  KEYWORD: 'keyword',
-  SCOPE: 'scope',
-};
-
 export const DEFAULT_PLAY_TRACK = 'spotify:album:5ht7ItJgpBH7W6vJ5BqpPr';
 
 export const SPOTIFY_WEB_API = new SpotifyWebApi();
