+++ src/constants/url.ts
@@ -1,0 +1,20 @@
+export const API = {
+  LOGIN: 'https://vinylify-express.vercel.app/',
+  SPOTIFY: 'https://api.spotify.com/v1/',
+  GENIUS:
+    '/genius/api' /** proxy vite.config.ts 설정 (https://api.genius.com) */,
+  EXPRESS: 'https://vinylify-express.vercel.app/',
+};
+export const PAGE = {
+  MAIN: '/',
+  ERROR: '/error',
+  MYPAGE: '/mypage',
+  SEARCH: '/search',
+  MUSIC_INFO: '/music-info',
+  LOGGED_IN: '/me',
+};
+
+export const URL_PARAMS = {
+  KEYWORD: 'keyword',
+  SCOPE: 'scope',
+};
